requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
paddlepaddle==2.5.2
paddleocr==2.7.0
reportlab
Pillow
numpy
